+++ changelog.md
@@ -1,10 +1,22 @@
 #### Version Stack
 ========================
 
-* Latest development version: `6.3.00-RELEASE` (branch `main`)
+* Latest development version: `6.3.01-RELEASE` (branch `main`)
 
 #### New Features and Changes
 ========================
+
+> :rocket: **Version 6.3.01-RELEASE**
+>
+> [BUG FIX]
+> * Fix: Driver Download Issue
+
+> [OPTIMIZATION]
+> * update pom versions to reduce vulnerability of dependencies
+> * update proxy factory of rest driver
+> * simplify log and screenshot file name
+
+
 
 > :rocket: **Version 6.3.00-RELEASE**
 >